CRAYSAT-1241: Remove dependency on kubectl
This commit removes the RPM requirement on kubectl, because it cannot
be satisfied as a part of the NCN image build process (see: MTL_1570).

In the spec file, comment out the requirement and add a reference to
MTL_1570. Update change log.

Test Description: After building the RPM, I temporarily modified
a local version of csm-rpms and built the k8s image using Packer.
Then, I chrooted into the image to check that prodmgr would at
least print its help text.