CASM-3971: Options to specify the deletion container and version
This adds to new optional arguments to the parser. They allow you
to specify the full path to the deletion container and also its
version. This allows the deletion container to be configurable in
case you wanted to use a custom one or an older version.

Add the ability to look up the version of the
product-deletion-utility container in the product catalog based on
its CSM version.

Find the latest version of product-deletion-utility

Add the ability to look up the version of the
product-deletion-utility container as part of the build. This
version is supplied as the default version in the argument parser.
Thus, if a version for that container is not supplied when prodmgr
is invoked, the default version is used.

Remove cms-meta-tools.spec file during runbuild prep, so the rpm
can be built successfully.